# [ matrix_transpose/bugfix ] Prevent reading/saving data from/to unallocated memory